fix a test sometimes failing on some machines
Due to timing issues, waitFor() sometimes gave up before the tooltip was
removed from the DOM after un-hovering the element. This made a test of the
timeline page fail intermittently. This commit seems to fix the issue, by
increasing the timeout of that specific waitFor() function from 1s (the
default) to 3s.

re #66